# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.SetPeriodeRequest:
    properties:
      BULAN:
        type: string
      TAHUN:
        type: string
    required:
    - BULAN
    - TAHUN
    type: object
...
  models.SDBNOMOR:
    properties:
      akm:
        type: string
      alias:
        type: string
      avt:
        type: string
      bbk:
        type: string
      bbm:
        type: string
      bbp:
        type: string
      bhn:
        type: string
      bjk:
        type: string
      bkk:
        description: Dynamic Numbering Fields
        type: string
      bkm:
        type: string
      bmm:
        type: string
      bp:
        type: string
      bpb:
        type: string
      bpl:
        type: string
      bpo:
        type: string
      bppb:
        type: string
      bpsb:
        type: string
      bpsprk:
        type: string
      bsprk:
        type: string
      contoh:
        type: string
      cr:
        type: string
      digitNomor:
        type: string
      dn:
        type: string
      enq:
        type: string
      format1:
        type: integer
      format2:
        type: integer
      format3:
        type: integer
      format4:
        type: integer
      hpd:
        type: string
      ht:
        type: string
      inicab:
        type: string
      ins:
        type: string
      inv:
        type: string
      invc:
        type: string
      kmbj:
        type: string
      kms:
        type: string
      kn:
        type: string
      kns:
        type: string
      krs:
        type: string
      noInvc:
        type: string
      noKn:
        type: string
      noSpr:
        type: string
      nobbk:
        type: string
      nobbm:
        type: string
      nobbp:
        type: string
      nobhn:
        type: string
      nobjk:
        type: string
      nobkk:
        type: string
      nobkm:
        type: string
      nobmm:
        type: string
      nobp:
        type: string
      nobpb:
        type: string
      nobpl:
        type: string
      nobpo:
        type: string
      nobppb:
        type: string
      nobpsb:
        type: string
      nobpsprk:
        type: string
      nobsprk:
        type: string
      nocr:
        type: string
      noenq:
        type: string
      noins:
        type: string
      noinv:
        type: string
      nokms:
        type: string
      nokns:
        type: string
      noopn:
        type: string
      nopbl:
        type: string
      nopjl:
        type: string
      nopnj:
        type: string
      nopo:
        type: string
      noppl:
        type: string
      norbj:
        type: string
      norbpb:
        type: string
      norkl:
        type: string
      norpb:
        type: string
      norpj:
        type: string
      norspb:
        type: string
      nosc:
        type: string
      noseri:
        type: string
      noship:
        type: string
      noso:
        type: string
      nospb:
        type: string
      nospg:
        type: string
      nospp:
        type: string
      nosprk:
        type: string
      notbj:
        type: string
      notrc:
        type: string
      notrs:
        type: string
      opbj:
        type: string
      opn:
        type: string
      pbl:
        type: string
      pbs:
        type: string
      pemisah:
        type: string
      pjl:
        type: string
      pnj:
        type: string
      po:
        type: string
      pos:
        type: string
      ppl:
        type: string
      pr:
        type: string
      pt:
        type: string
      rbj:
        type: string
      rbp:
        type: string
      rbpb:
        type: string
      reset:
        type: integer
      rkl:
        type: string
      rpb:
        type: string
      rpbs:
        type: string
      rpj:
        type: string
      rspb:
        type: string
      sc:
        type: string
      ship:
        type: string
      so:
        type: string
      spb:
        type: string
      spg:
        type: string
      spk:
        description: Other string codes
        type: string
      spp:
        type: string
      spr:
        type: string
      sprk:
        type: string
      tbj:
        type: string
      trc:
        type: string
      trs:
        type: string
      tt:
        type: string
    type: object
  models.SDBPERUSAHAAN:
    properties:
      alamat1:
        type: string
      alamat2:
        type: string
      alamatpkp1:
        type: string
      alamatpkp2:
        type: string
      alamatpkp21:
        type: string
      alamatpkp22:
        type: string
      direksi:
        type: string
      email:
        type: string
      fax:
        type: string
      jabatan:
        type: string
      kodeusaha:
        type: string
      kota:
        type: string
      kotapkp:
        type: string
      kotapkp1:
        type: string
      nama:
        type: string
      namapkp:
        type: string
      namapkp1:
        type: string
      npwp:
        type: string
      npwp1:
        type: string
      telpon:
        type: string
      tglpengukuhan:
        type: string
      tglpengukuhan1:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for DAPEN management of employee retirement funds
  title: DAPEN System API
  version: "1.0"
paths:
...
  /api/admin/settings/company:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SDBPERUSAHAAN'
      summary: Get Company Settings
      tags:
      - Admin Settings
    put:
      consumes:
      - application/json
      parameters:
      - description: Company Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SDBPERUSAHAAN'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update Company Settings
      tags:
      - Admin Settings
  /api/admin/settings/numbers:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SDBNOMOR'
      summary: Get Numbering Settings
      tags:
      - Admin Settings
    put:
      consumes:
      - application/json
      parameters:
      - description: Numbering Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SDBNOMOR'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update Numbering Settings
      tags:
      - Admin Settings
...
  /api/admin/users/{userId}/sessions:
    delete:
      consumes:
      - application/json
      description: Revoke all active sessions for a user. The user will be logged
        out from all devices.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Revoke All User Sessions
      tags:
      - Session Management
    get:
      consumes:
      - application/json
      description: Fetch all active sessions for a specific user. Admin-only endpoint.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List User Active Sessions
      tags:
      - Session Management
  /api/admin/users/{userId}/sessions/{sessionId}:
    delete:
      consumes:
      - application/json
      description: Revoke a specific session by sessionId. The user will be logged
        out from that device/session.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Revoke a Single Session
      tags:
      - Session Management
...
  /berkas/get-periode:
    get:
      description: Fetch the active periode for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Current Periode
      tags:
      - Berkas
  /berkas/set-periode:
    put:
      consumes:
      - application/json
      description: Update the active periode for the user
      parameters:
      - description: Periode Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SetPeriodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Set Current Periode
      tags:
      - Berkas
...
swagger: "2.0"